Move the agents directory from a location in Application Support to a temporary directory.
In macOS Catalina, all directories in Application Support are managed by TCC. If TCC thinks that an application is modifying a directory in Application Support which is not its own, it will deny any change to that directory even if filesystem permissions are correct.

By creating a different directory for each agent, and by locating this directories inside $TMPDIR, we avoid triggering TCC's wrath. launchd doesn't care if we launch agents from a temporary directory.

Other improvements introduced as side-effects:
(1) the agents directory is now removed completely as soon as there are no agents running
(2) the agents directory will always be located in the root drive (because all temporary directories are created under /private/var/folders), therefore improving performance if the user's home is remapped to a slower disk